src/requirements.txt
pymysql
mysqlclient
mysql-connector-python